requirements.txt

diffusers
transformers
pillow
accelerate
peft
streamlit
opencv-python
streamlit_image_select
omegaconf
einops
streamlit_drawable_canvas

